+++ src/blueprint/agents/agent/__init__.py
@@ -1,31 +1,14 @@
 from .agent_builder import AgentBuilder
-from .agent_factory import AgentFactory
 from .agent_runtime import AgentRuntime
+from .metrics import MetricsExtractor, MetricsRecorder
 from .model_provider import ModelProviderFactory
 from .prompt_loader import PromptLoader
-from .providers import (
-    OpenAIAgentFactory,
-    OpenAIModelProvider,
-    OpenAIResponseHandler,
-    VLLMAgentFactory,
-    VLLMModelProvider,
-    VLLMResponseHandler,
-)
-from .response_handler import ResponseHandlerFactory
-from .usage_limits import UsageLimitsBuilder
 
 __all__ = [
     "AgentBuilder",
-    "AgentFactory",
     "AgentRuntime",
     "ModelProviderFactory",
-    "ResponseHandlerFactory",
     "PromptLoader",
-    "UsageLimitsBuilder",
-    "OpenAIModelProvider",
-    "OpenAIResponseHandler",
-    "OpenAIAgentFactory",
-    "VLLMModelProvider",
-    "VLLMResponseHandler",
-    "VLLMAgentFactory",
+    "MetricsRecorder",
+    "MetricsExtractor",
 ]